+++ movies-explorer-frontend/src/components/Navigation/Navigation.css
@@ -4,14 +4,14 @@
     position: fixed;
     width: 100%;
     height: 100%;
-    background-color: rgba(0, 0, 0, 0.9);
+    background-color: rgba(0, 0, 0, 0.3);
     align-items: flex-end;
     justify-content: center;
     visibility: hidden;
     opacity: 0;
     top: 0;
     left: 0;
-    transition: visibility 0s linear 0.5s, opacity 0.5s linear;
+    transition: visibility 0s linear 0.3s, opacity 0.3s linear;
 }
 
 .navigation_is-opened {
@@ -19,6 +19,7 @@
     visibility: visible;
     opacity: 1;
     transition-delay: 0s;
+    z-index: 1;
 }
 
 .navigation__container {
@@ -47,6 +48,71 @@
 }
 
 .navigation__main-link {
-
+    display: flex;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 18px;
+    line-height: 22px;
+    color: #000000;
+    text-decoration: none;
+    margin: 0;
+    margin-top: 105.44px;
 }
 
+.navigation__login-button {
+    display: flex;
+    width: 100px;
+    height: 32px;
+    background-color: #FFFFFF;
+    border-radius: 40px;
+    box-shadow: 0px 3px 14px rgba(0, 0, 0, 0.08);
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 621px;
+}
+
+.navigation__text {
+    font-size: 13px;
+    line-height: 12px;
+    color: #000000;
+    margin: 0;
+    margin-left: 11px;
+}
+
+.navigation__profile-wrapper {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 26px;
+    height: 26px;
+    background: #F9F9F9;
+    border-radius: 50px;
+    margin-right: 3px;
+}
+
+.navigation__profile-logo {
+    background-image: url(../../images/icon__profile-logo.svg);
+    background-repeat: no-repeat;
+    background-size: contain;
+    width: 10.38px;
+    height: 10.67px;  
+}
+
+@media screen and (max-width: 320px) {
+    .navigation__container {
+        max-width: 320px;
+    }
+    
+    .navigation__close-btn {
+        margin-top: 13.44px;
+        margin-right: 13.44px;
+    }
+    
+    .navigation__main-link {
+        margin-top: 99.44px;
+    }
+    
+    .navigation__login-button {
+        margin-top: 444px;
+    }
+}
