+++ movies-explorer-frontend/src/components/ProtectedRoute/ProtectedRoute.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Route, Redirect } from "react-router-dom";
 import Header from "../Header/Header";
+import Navigation from '../Navigation/Navigation';
 
 const ProtectedRoute = ({component: Component, ...props}) => {
   return (
@@ -8,9 +9,6 @@ const ProtectedRoute = ({component: Component, ...props}) => {
         <Header
             loggedIn={props.loggedIn}
             path={props.path}
-            onNavBtnClick={props.onNavBtnClick}
-            isNavigationOpen={props.isNavigationOpen}
-            closeAllPopups={props.closeAllPopups}
         />
         <Route path={props.path}>
         {
